Two-linked list with templates
