--- README.md
@@ -1,2 +1,0 @@
-# DataStructures
-Репозиторий с реализацией структур данных (Krage56)
